refactor(glm5): remove retired fast-path helpers
Ultraworked with [omo](https://github.com/code-yeongyu/oh-my-openagent)
